refactor: address clean-code findings
